+++ src/main/java/roomescape/controller/ReservationController.java
@@ -3,6 +3,7 @@ package roomescape.controller;
 import jakarta.validation.Valid;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import roomescape.controller.dto.ReservationCreateRequest;
 import roomescape.entity.Reservation;
 import roomescape.service.ReservationService;
