+++ src/main/java/roomescape/controller/dto/ReservationTimeCreateRequest.java
@@ -1,4 +1,5 @@
 package roomescape.controller;
+package roomescape.controller.dto;
 
 import java.time.LocalTime;
